// src/Types/ProductDetails.ts
export type ProductDetails = {
  id: string;
  name: string;
  capacityAvailable: string[];
  capacity: string;
  priceRegular: number;
  priceDiscount: number;
  colorsAvailable: string[];
  color: string;
  images: string[];
  resolution: string;
  processor: string;
  ram: string;
  camera: string;
  zoom: string;
  cell: string[];
  namespaceId: string;
  category: string;
  screen: string;
  description: Description[];
};

type Description = {
  title: string;
  text: string;
};